requirements.txt
appnope==0.1.2
argcomplete==1.12.2
argh==0.26.2
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
backcall==0.2.0
biopython @ file:///opt/concourse/worker/volumes/live/cdcffdf0-9e67-4291-4fd7-8d3bf14bf437/volume/biopython_1600298878668/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///opt/concourse/worker/volumes/live/8fbed55e-0573-4246-7cf9-162067141d5e/volume/cffi_1605538070636/work
chardet @ file:///opt/concourse/worker/volumes/live/a5a7c56b-23cb-471d-7404-f4130b8aed33/volume/chardet_1605303184529/work
conda==4.9.2
conda-package-handling @ file:///opt/concourse/worker/volumes/live/a7e34989-4c54-4cb6-4156-4e58ee270730/volume/conda-package-handling_1603018121300/work
cryptography @ file:///opt/concourse/worker/volumes/live/62f9527a-361c-47e5-5709-227a4a523add/volume/cryptography_1605544472602/work
decorator==4.4.2
gffutils==0.10.1
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.20.0
olefile==0.46
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///opt/concourse/worker/volumes/live/705be880-f90c-4fb3-489d-9ee685c52cbe/volume/pillow_1609786792428/work
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pycodestyle==2.6.0
pycollada==0.6
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyfaidx==0.5.9.2
Pygments==2.7.4
pykerberos==1.2.1
pymol==2.4.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pyzmq==22.0.2
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
simplejson==3.17.2
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.0.5
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth==0.2.5
